Issue #18443: Move viewholders to separate files and package
