+++ app/src/main/java/org/mozilla/fenix/tabstray/viewholders/AbstractTrayViewHolder.kt
@@ -1,0 +1,23 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package org.mozilla.fenix.tabstray.viewholders
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+import kotlinx.android.extensions.LayoutContainer
+import org.mozilla.fenix.tabstray.TrayPagerAdapter
+
+/**
+ * An abstract [RecyclerView.ViewHolder] for [TrayPagerAdapter] items.
+ */
+abstract class AbstractTrayViewHolder constructor(
+    override val containerView: View
+) : RecyclerView.ViewHolder(containerView), LayoutContainer {
+
+    abstract fun bind(
+        adapter: RecyclerView.Adapter<out RecyclerView.ViewHolder>,
+        layoutManager: RecyclerView.LayoutManager
+    )
+}
