+++ app/src/main/java/org/mozilla/fenix/tabstray/viewholders/NormalBrowserTabViewHolder.kt
@@ -2,59 +2,26 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.tabstray
+package org.mozilla.fenix.tabstray.viewholders
 
 import android.view.View
-import androidx.annotation.CallSuper
 import androidx.recyclerview.selection.SelectionPredicates
 import androidx.recyclerview.selection.SelectionTracker
-import androidx.recyclerview.selection.StableIdKeyProvider
 import androidx.recyclerview.selection.StorageStrategy
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.extensions.LayoutContainer
 import org.mozilla.fenix.R
-import org.mozilla.fenix.tabstray.browser.BaseBrowserTrayList
+import org.mozilla.fenix.tabstray.TabsTrayInteractor
 import org.mozilla.fenix.tabstray.browser.BrowserTabsAdapter
 import org.mozilla.fenix.tabstray.browser.TabsDetailsLookup
+import org.mozilla.fenix.tabstray.browser.TabsItemKeyProvider
 
 /**
- * Base [RecyclerView.ViewHolder] for [TrayPagerAdapter] items.
+ * View holder for the normal tabs tray list.
  */
-abstract class TrayViewHolder constructor(
-    override val containerView: View
-) : RecyclerView.ViewHolder(containerView), LayoutContainer {
-
-    abstract fun bind(
-        adapter: RecyclerView.Adapter<out RecyclerView.ViewHolder>,
-        layoutManager: RecyclerView.LayoutManager
-    )
-}
-
-abstract class BrowserTabViewHolder(
-    containerView: View,
-    interactor: TabsTrayInteractor
-) : TrayViewHolder(containerView) {
-
-    protected val trayList: BaseBrowserTrayList = itemView.findViewById(R.id.tray_list_item)
-
-    init {
-        trayList.interactor = interactor
-    }
-
-    @CallSuper
-    override fun bind(
-        adapter: RecyclerView.Adapter<out RecyclerView.ViewHolder>,
-        layoutManager: RecyclerView.LayoutManager
-    ) {
-        trayList.layoutManager = layoutManager
-        trayList.adapter = adapter
-    }
-}
-
 class NormalBrowserTabViewHolder(
     containerView: View,
     interactor: TabsTrayInteractor
-) : BrowserTabViewHolder(containerView, interactor) {
+) : BaseBrowserTabViewHolder(containerView, interactor) {
 
     private lateinit var selectionTracker: SelectionTracker<Long>
 
@@ -85,14 +52,5 @@ class NormalBrowserTabViewHolder(
 
     companion object {
         const val LAYOUT_ID = R.layout.normal_browser_tray_list
-    }
-}
-
-class PrivateBrowserTabViewHolder(
-    containerView: View,
-    interactor: TabsTrayInteractor
-) : BrowserTabViewHolder(containerView, interactor) {
-    companion object {
-        const val LAYOUT_ID = R.layout.private_browser_tray_list
     }
 }
